fix docs page
