+++ .tina/schema.tsx
@@ -12,18 +12,19 @@ limitations under the License.
 */
 
 import { defineSchema } from 'tinacms'
-import { heroTemplate } from '../components/blocks/Hero'
-import { featuresTemplate } from '../components/blocks/Features'
-import { flyingTemplate } from '../components/blocks/Flying'
-import { pricingTemplate } from '../components/blocks/Pricing'
-import { faqTemplate } from '../components/blocks/FAQ'
-import { contentTemplate } from '../components/blocks/Content'
-import { columnsTemplate } from '../components/blocks/Columns'
-import { showcaseTemplate } from '../components/blocks/Showcase'
-import { storyTemplate } from '../components/blocks/Story'
-import { featureGridTemplate } from '../components/blocks/FeatureGrid'
 import type { TinaTemplate } from 'tinacms'
-import { logoGridTemplate } from 'components/blocks'
+
+import { heroTemplate } from '../components/blocks/Hero.template'
+import { featuresTemplate } from '../components/blocks/Features.template'
+import { flyingTemplate } from '../components/blocks/Flying.template'
+import { pricingTemplate } from '../components/blocks/Pricing.template'
+import { faqTemplate } from '../components/blocks/FAQ.template'
+import { contentTemplate } from '../components/blocks/Content.template'
+import { columnsTemplate } from '../components/blocks/Columns.template'
+import { showcaseTemplate } from '../components/blocks/Showcase.template'
+import { storyTemplate } from '../components/blocks/Story.template'
+import { featureGridTemplate } from '../components/blocks/FeatureGrid.template'
+import { logoGridTemplate } from '../components/blocks/LogoGrid.template'
 
 export default defineSchema({
   config: {
