+++ components/blocks/Actions.tsx
@@ -4,77 +4,6 @@ import { IconRight } from './Icons'
 import { copyToClipboard } from 'components/layout/MarkdownContent'
 import { LinkButton } from 'components/ui'
 
-export const actionsTemplate = {
-  label: 'Actions',
-  name: 'actions',
-  type: 'object',
-  list: true,
-  ui: {
-    defaultItem: {
-      variant: 'default',
-      label: 'Secondary Action',
-      icon: false,
-      size: 'medium',
-      url: '/',
-    },
-  },
-  fields: [
-    { name: 'label', label: 'Label', type: 'string' },
-    { name: 'icon', label: 'Icon', type: 'boolean' },
-    {
-      name: 'variant',
-      label: 'Variant',
-      type: 'string',
-      options: [
-        {
-          value: 'default',
-          label: 'Seafoam',
-        },
-        {
-          value: 'blue',
-          label: 'Blue',
-        },
-        {
-          value: 'orange',
-          label: 'Orange',
-        },
-        {
-          value: 'white',
-          label: 'White',
-        },
-        {
-          value: 'ghost',
-          label: 'Ghost',
-        },
-        {
-          value: 'command',
-          label: 'Command',
-        },
-      ],
-    },
-    {
-      name: 'size',
-      label: 'Size',
-      type: 'string',
-      options: [
-        {
-          value: 'small',
-          label: 'Small',
-        },
-        {
-          value: 'medium',
-          label: 'Medium',
-        },
-        {
-          value: 'large',
-          label: 'Large',
-        },
-      ],
-    },
-    { name: 'url', label: 'URL', type: 'string' },
-  ],
-}
-
 export const Actions = ({ items, align = 'left' }) => {
   return (
     <>
