+++ components/blocks/Flying.tsx
@@ -1,26 +1,6 @@
-import { actionsTemplate, Actions } from './Actions'
+import { Actions } from './Actions'
 import CloudsOne from '../../public/svg/clouds-1.svg'
 import CloudsTwo from '../../public/svg/clouds-2.svg'
-import type { TinaTemplate } from '@tinacms/cli'
-
-export const flyingTemplate: TinaTemplate = {
-  name: 'flying',
-  label: 'Flying',
-  ui: {
-    previewSrc: '/img/blocks/flying.png',
-  },
-  fields: [
-    { name: 'headline', label: 'Headline', type: 'string' },
-    {
-      name: 'text',
-      label: 'Text',
-      ui: { component: 'textarea' },
-      type: 'string',
-    },
-    // @ts-ignore
-    actionsTemplate,
-  ],
-}
 
 export function FlyingBlock({ data, index }) {
   return (
