+++ components/blocks/Hero.tsx
@@ -1,42 +1,7 @@
 import React from 'react'
-import { actionsTemplate, Actions } from './Actions'
+import { Actions } from './Actions'
 import { Container } from './Container'
-import HeroBackground from '../../public/svg/hero-bg.svg'
-import type { TinaTemplate } from '@tinacms/cli'
 import HeroBg from '../../public/svg/hero-bg.svg'
-
-export const heroTemplate: TinaTemplate = {
-  label: 'Hero',
-  name: 'hero',
-  ui: {
-    previewSrc: '/img/blocks/hero.png',
-    defaultItem: {
-      headline: 'Next Gen Content Management',
-      text: 'Tina is an open-source, Git-backed CMS with the ability to add visual editing to your NextJS site',
-      actions: [
-        {
-          variant: 'orange',
-          label: 'Primary Action',
-          icon: true,
-          url: '/',
-        },
-        {
-          variant: '',
-          label: 'Secondary Action',
-          icon: false,
-          url: '/',
-        },
-      ],
-    },
-  },
-  fields: [
-    { name: 'headline', label: 'Headline', type: 'string' },
-    { name: 'text', label: 'Text', type: 'string' },
-    // @ts-ignore
-    actionsTemplate,
-    { name: 'videoSrc', label: 'Video Source', type: 'string' },
-  ],
-}
 
 export function HeroBlock({ data, index }) {
   return (
