+++ components/blocks/Social.tsx
@@ -4,42 +4,7 @@ import { FaDiscord, FaTwitter, FaGithub } from 'react-icons/fa'
 import TinaIconSvg from '../../public/svg/tina-icon.svg'
 import { Button, ButtonGroup } from '../ui'
 
-export const socialTemplate = {
-  name: 'social',
-  label: 'Social Links',
-  ui: {
-    defaultItem: {
-      tina: 'https://github.com/tinacms/tinacms/discussions',
-      discord: 'https://discord.com/invite/zumN63Ybpf',
-      github: 'https://github.com/tinacms/tinacms',
-      twitter: 'https://twitter.com/tinacms',
-    },
-  },
-  fields: [
-    {
-      type: 'string',
-      name: 'tina',
-      label: 'Tina',
-    },
-    {
-      type: 'string',
-      name: 'discord',
-      label: 'Discord',
-    },
-    {
-      type: 'string',
-      name: 'github',
-      label: 'Github',
-    },
-    {
-      type: 'string',
-      name: 'twitter',
-      label: 'Twitter',
-    },
-  ],
-}
-
-export const SocialBlock = props => {
+export const SocialBlock = (props) => {
   return (
     <ButtonGroup>
       {props.tina && (
