+++ components/blocks/Story.template.ts
@@ -1,0 +1,13 @@
+import type { TinaTemplate } from 'tinacms'
+
+export const storyTemplate: TinaTemplate = {
+  label: 'Story',
+  name: 'story',
+  fields: [
+    {
+      label: 'Title',
+      name: 'title',
+      type: 'string',
+    },
+  ],
+}
