+++ package.json
@@ -70,6 +70,7 @@
     "nextjs-cors": "^2.1.0",
     "node-fetch": "^2.6.1",
     "node-sass": "^7.0.0",
+    "parse-numeric-range": "^1.3.0",
     "polished": "^3.4.4",
     "postcss": "^8.4.18",
     "prop-types": "^15.7.2",
@@ -87,10 +88,12 @@
     "react-instantsearch-dom": "^6.3.0",
     "react-intersection-observer": "^9.4.0",
     "react-live": "^2.3.0",
-    "react-markdown": "^5.0.2",
+    "react-markdown": "^8.0.3",
     "react-syntax-highlighter": "^15.5.0",
     "react-textarea-autosize": "^7.1.2",
     "react-tinacms-editor": "^0.53.9",
+    "rehype-raw": "^6.1.1",
+    "rehype-sanitize": "^5.0.1",
     "remark": "^11.0.2",
     "remark-parse": "^8.0.0",
     "remark-shortcodes": "^0.3.1",
